+++ bridge/types.go
@@ -28,6 +28,7 @@ type Config struct {
 	RefreshInterval int
 	DeregisterCheck string
 	Cleanup         bool
+	HostnameTag	bool
 }
 
 type Service struct {